fix : Gemini's feedback
