+++ wise-office-client/src/components/modal/ProjectModal/ProjectModal.tsx
@@ -12,6 +12,7 @@ import LoadingIndicator from "@/components/ui/LoadingIndicator";
 import { projectDefaultForm } from "@/constants/defaultForm";
 import { projectFormValidate } from "@/utils/formValidation";
 import { toastMessage } from "@/lib/common/toastMessage";
+import { useScrollLock } from "@/hooks/useBodyScrollLock";
 
 type ProjectModalProps = {
     projectId?: number;
@@ -57,14 +58,7 @@ export default function ProjectModal({
         });
     }, [data]);
 
-    useEffect(() => {
-        const originalStyle = window.getComputedStyle(document.body).overflow;
-        document.body.style.overflow = "hidden";
-
-        return () => {
-            document.body.style.overflow = originalStyle;
-        };
-    }, []);
+    useScrollLock();
 
     /* ----- func -----*/
     const handleChange = <K extends keyof CreateProject>(
